Moving http files under sse folder to avoid conflicts when the lua-resty-http module is required. require('resty.http') was bringing the lua-resty-sse version vs the lua-resty-http version
